refactor([issue-747]): split voice tools.js into per-domain modules
server/services/voice/tools.js had grown to ~2200 lines holding 41 tool
definitions plus all their helpers and intent regexes. Move each tool's
schema+execute into a per-domain module under server/services/voice/tools/
(brain, meatspace, goals, system, dailylog, ui, ask, vision, ambient,
timer, media, pipeline, code, catalog) with a shared.js for the cross-domain
helpers (clampLimit, findUiElement, UI_KINDS).

tools.js stays the single source of truth for action schemas: it imports
each domain's tool array + intent regex, assembles the TOOLS registry and
the TOOL_GROUPS/GROUP_INTENT maps, and re-exports the unchanged public API
(getToolSpecs, getToolSpecsForIntent, getAllToolNames, getToolMetadata,
classifyIntent, dispatchTool, UI_KINDS, UI_INTENT_RE, UI_FILL_INTENT_RE,
anchorLocalMidnightUtc). The palette route and voice pipeline import sites
are untouched. Added a second boot-time guard that every TOOL_GROUPS key
references a registered tool, complementing the existing group-has-regex
guard.
